requirements.txt
# DermaAI - Skin Disease Classification System
# Requirements

# ── Web Framework ────────────────────────────────────────
streamlit>=1.32.0

# ── Deep Learning ────────────────────────────────────────
tensorflow>=2.13.0
keras>=2.13.0

# ── Machine Learning ─────────────────────────────────────
scikit-learn>=1.3.0
joblib>=1.3.0

# ── Computer Vision ──────────────────────────────────────
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ── Numerical Computing ──────────────────────────────────
numpy>=1.24.0

# ── Utilities ────────────────────────────────────────────
python-dotenv>=1.0.0

opencv-python
